requirements.txt
numpy
opencv-python
pyquaternion
nuscenes-devkit
Pillow
requests
tqdm